# rubocop: reduce a regex match for performance