@kenhys 2025-04-27
NOTE:

rubocop detects the following code, but it was a false positive.

```
Ofenses:

lib/fluent/command/plugin_config_formatter.rb:64:17: C: [Correctable] Performance/StringInclude: Use String#include? instead of a regex match with literal-only pattern.
        next if /::PluginHelper::/.match?(plugin_class.name)
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```

```
ruby 3.2.8 (2025-03-26 revision 13f495dc2c) +YJIT [x86_64-linux]
Warming up --------------------------------------
check method with match?
                         1.291M i/100ms
check method with include?
                       744.908k i/100ms
Calculating -------------------------------------
check method with match?
                         13.453M (± 0.7%) i/s   (74.33 ns/i) -     68.399M in   5.084622s
check method with include?
                          7.220M (± 1.0%) i/s  (138.50 ns/i) -     36.500M in   5.055801s

Comparison:
check method with match?: 13452792.6 i/s
check method with include?:  7220336.0 i/s - 1.86x  slower
```